refactor: split game, map, ui
